feat(project): Initial import
I've added a docker-compose.yaml file to run a dockerized version of pgadmin4
that can access postgres databases on other docker networks, this is important
as the databases will be part of individual apps but this pgadmin is outside
those docker networks.

I have also added a `servers.json` file that will record all connection parameters
outside the password, so that devs can share this instance and have essentially a
zero-effort access to local/remote databases.